refactor: prepare uuid extension
